chore: properties refactoring
